# docs/swagger.yaml
basePath: /v1
definitions:
...
  swagger.CoachtechData:
    properties:
      example:
        example: example_data
        type: string
      example1:
        example: example_data_1
        type: string
      example2:
        example: example_data_2
        type: string
      example3:
        example: example_data_3
        type: string
      example4:
        example: example_data_4
        type: string
    type: object
  swagger.CoachtechInsertInput:
    properties:
      coachtech_id:
        example: 123
        type: integer
      data:
        $ref: '#/definitions/swagger.CoachtechData'
      summary_date:
        example: "2025-07-14"
        type: string
      test_id:
        example: endurance_test_1
        type: string
      user_id:
        example: 1c2f6ad2-dc8c-4c44-85e2-381d70b093ef
        type: string
    type: object
  swagger.CoachtechStatusResponse:
    properties:
      data:
        example: true
        type: boolean
    type: object
...
  swagger.ForbiddenError:
    properties:
      error:
        example: forbidden
        type: string
    type: object
  swagger.ForbiddenResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ForbiddenError'
        type: array
    type: object
...
  swagger.InternalServerError:
    properties:
      error:
        example: the server encountered a problem
        type: string
    type: object
  swagger.InternalServerErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.InternalServerError'
        type: array
    type: object
...
  swagger.ServiceUnavailableError:
    properties:
      error:
        example: database is unavailable
        type: string
    type: object
  swagger.ServiceUnavailableResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ServiceUnavailableError'
        type: array
    type: object
...
  swagger.UnauthorizedError:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
  swagger.UnauthorizedResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.UnauthorizedError'
        type: array
    type: object
...
  swagger.ValidationError:
    properties:
      parameter name:
        example: full description of the error
        type: string
    type: object
  swagger.ValidationErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ValidationError'
        type: array
    type: object
info:
  contact: {}
  description: API for integrating, analyzing, and visualizing sports and health data
  termsOfService: https://csc.fi/en/security-privacy-data-policy-and-open-source-policy/privacy/
  title: KUHA REST API
paths:
...
  /utv/coachtech/data:
    get:
      consumes:
      - application/json
      description: Returns Coachtech data for a user
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Filter data after this date (YYYY-MM-DD)
        in: query
        name: after_date
        type: string
      - description: Filter data before this date (YYYY-MM-DD)
        in: query
        name: before_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Get Coachtech data
      tags:
      - UTV - Coachtech
  /utv/coachtech/insert:
    post:
      consumes:
      - application/json
      description: Adds a Coachtech ID and corresponding data for a user
      parameters:
      - description: Coachtech data input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swagger.CoachtechInsertInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Insert Coachtech data
      tags:
      - UTV - Coachtech
  /utv/coachtech/status:
    get:
      consumes:
      - application/json
      description: Returns whether Coachtech data exists for a given user
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.CoachtechStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Check Coachtech data availability
      tags:
      - UTV - Coachtech
...
securityDefinitions:
  BearerAuth:
    description: 'Use format: Bearer your_JWT_here'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"